requirements.txt
crewai==0.108.0
openai==1.68.0
python-dotenv==1.0.1
Requests==2.32.3
